refactor(core): centralize fetchTokenDomain + EIP-5267/name/version ABIs
The facilitator's verify path (`fetchTokenDomain`) and the e2e harness's
`createChainTokenDomainResolver` independently re-declared `EIP5267_ABI`,
`NAME_ABI`, `VERSION_ABI` and the same EIP-5267 → ERC-20 `name()` +
EIP-2612 `version()` fallback flow. Two copies of the signing-domain
lookup risks signing/verification drift over time.

Lift the ABIs and the `fetchTokenDomain` function into
`@bosonprotocol/x402-core/eip712/token-auth/fetch-token-domain.ts` and
re-export them from the existing `eip712/token-auth` barrel. The
facilitator and the harness now both consume the same primitive, so
they're guaranteed to read the same domain from the same ABI surface.

Addresses PR #88 review comment r3283386286.
